+++ src/main/java/dev/willyelton/crystal_tools/client/gui/CrystalGeneratorScreen.java
@@ -2,18 +2,19 @@ package dev.willyelton.crystal_tools.client.gui;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import dev.willyelton.crystal_tools.CrystalTools;
-import dev.willyelton.crystal_tools.client.gui.component.EnergyBarWidget;
 import dev.willyelton.crystal_tools.client.gui.component.BlockEntityUpgradeButton;
-import dev.willyelton.crystal_tools.common.inventory.container.CrystalGeneratorContainerMenu;
+import dev.willyelton.crystal_tools.client.gui.component.EnergyBarWidget;
+import dev.willyelton.crystal_tools.common.inventory.container.AbstractGeneratorContainerMenu;
 import dev.willyelton.crystal_tools.utils.IntegerUtils;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 
 // TODO: Superclass for this and furnace could be useful. Drawing fire + animations + upgrade button + skill point bar, abstract the "bar" component?
-public class CrystalGeneratorScreen extends AbstractContainerScreen<CrystalGeneratorContainerMenu> {
+public class CrystalGeneratorScreen extends AbstractContainerScreen<AbstractGeneratorContainerMenu> {
     private static final ResourceLocation TEXTURE = ResourceLocation.fromNamespaceAndPath(CrystalTools.MODID ,"textures/gui/crystal_generator.png");
 
     private static final int ENERGY_X = 8;
@@ -29,14 +30,14 @@ public class CrystalGeneratorScreen extends AbstractContainerScreen<CrystalGener
     private static final int FIRE_TEXTURE_OFF_X = 190;
     private static final int FIRE_TEXTURE_Y = 0;
 
-    private static final int UPGRADE_BUTTON_X = 155;
-    private static final int UPGRADE_BUTTON_Y = 6;
-    private static final int UPGRADE_BUTTON_WIDTH = 12;
-    private static final int UPGRADE_BUTTON_HEIGHT = 12;
+    protected static final int UPGRADE_BUTTON_X = 155;
+    protected static final int UPGRADE_BUTTON_Y = 6;
+    protected static final int UPGRADE_BUTTON_WIDTH = 12;
+    protected static final int UPGRADE_BUTTON_HEIGHT = 12;
 
     private final float expLabelX;
 
-    public CrystalGeneratorScreen(CrystalGeneratorContainerMenu container, Inventory inventory, Component title) {
+    public CrystalGeneratorScreen(AbstractGeneratorContainerMenu container, Inventory inventory, Component title) {
         super(container, inventory, title);
 
         this.imageHeight = 191;
@@ -47,8 +48,8 @@ public class CrystalGeneratorScreen extends AbstractContainerScreen<CrystalGener
     @Override
     protected void renderBg(GuiGraphics guiGraphics, float partialTick, int mouseX, int mouseY) {
         RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
-        RenderSystem.setShaderTexture(0, TEXTURE);
-        guiGraphics.blit(TEXTURE, leftPos, topPos, 0, 0, this.imageWidth, this.imageHeight);
+        RenderSystem.setShaderTexture(0, texture());
+        guiGraphics.blit(texture(), leftPos, topPos, 0, 0, this.imageWidth, this.imageHeight);
         this.renderFire(guiGraphics);
     }
 
@@ -84,7 +85,14 @@ public class CrystalGeneratorScreen extends AbstractContainerScreen<CrystalGener
                         UPGRADE_BUTTON_WIDTH,
                         UPGRADE_BUTTON_HEIGHT,
                         Component.literal("+"),
-                        pButton -> ModGUIs.openScreen(new BlockEntityUpgradeScreen(this.menu, this.menu.getPlayer(), this)),
+                        pButton -> {
+                            Screen toOpen = upgradeButtonScreen();
+                            if (toOpen == null) {
+                                CrystalTools.LOGGER.warn("Couldn't open upgrade button, no skill tree defined");
+                            } else {
+                                ModGUIs.openScreen(toOpen);
+                            }
+                        },
                         (button, guiGraphics, mouseX, mouseY) -> {
                             Component textComponent = Component.literal(this.menu.getSkillPoints() + " Point(s) Available");
                             guiGraphics.renderTooltip(this.font, this.font.split(textComponent, Math.max(CrystalGeneratorScreen.this.width / 2 - 43, 170)), mouseX, mouseY);
@@ -96,13 +104,29 @@ public class CrystalGeneratorScreen extends AbstractContainerScreen<CrystalGener
         );
     }
 
+    protected Screen upgradeButtonScreen() {
+        return new BlockEntityUpgradeScreen(this.menu, this.menu.getPlayer(), this);
+    }
+
+    protected ResourceLocation texture() {
+        return TEXTURE;
+    }
+
+    protected int fireX() {
+        return FIRE_X;
+    }
+
+    protected int fireY() {
+        return FIRE_Y;
+    }
+
     private void renderFire(GuiGraphics guiGraphics) {
-        guiGraphics.blit(TEXTURE, this.leftPos + FIRE_X, this.topPos + FIRE_Y, FIRE_TEXTURE_OFF_X, FIRE_TEXTURE_Y, FIRE_WIDTH, FIRE_HEIGHT);
+        guiGraphics.blit(TEXTURE, this.leftPos + fireX(), this.topPos + fireY(), FIRE_TEXTURE_OFF_X, FIRE_TEXTURE_Y, FIRE_WIDTH, FIRE_HEIGHT);
 
         if (this.menu.isLit()) {
             float litProgress = this.menu.getLitProgress();
             int height = (int) (litProgress * FIRE_HEIGHT);
-            guiGraphics.blit(TEXTURE, this.leftPos + FIRE_X, this.topPos + FIRE_Y + FIRE_HEIGHT - height, FIRE_TEXTURE_ON_X, FIRE_TEXTURE_Y + FIRE_HEIGHT - height, FIRE_WIDTH, height);
+            guiGraphics.blit(TEXTURE, this.leftPos + fireX(), this.topPos + fireY() + FIRE_HEIGHT - height, FIRE_TEXTURE_ON_X, FIRE_TEXTURE_Y + FIRE_HEIGHT - height, FIRE_WIDTH, height);
         }
     }
 }