+++ src/main/java/dev/willyelton/crystal_tools/common/components/DataComponents.java
@@ -2,6 +2,7 @@ package dev.willyelton.crystal_tools.common.components;
 
 import com.mojang.serialization.Codec;
 import dev.willyelton.crystal_tools.CrystalTools;
+import dev.willyelton.crystal_tools.common.datamap.GeneratorFuelData;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.core.registries.Registries;
@@ -185,6 +186,18 @@ public class DataComponents {
     public static final DeferredHolder<DataComponentType<?>, DataComponentType<Boolean>> INSTANT_PICKUP = register("instant_pickup", Codec.BOOL, ByteBufCodecs.BOOL, SkillType.BOOLEAN);
     public static final DeferredHolder<DataComponentType<?>, DataComponentType<Boolean>> DISABLE_MOB_PULL = register("disable_mob_pull", Codec.BOOL, ByteBufCodecs.BOOL, SkillType.BOOLEAN);
 
+    // Portable Generator
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Integer>> FE_GENERATION = register("fe_generation", Codec.INT, ByteBufCodecs.INT, SkillType.INT);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Float>> FUEL_EFFICIENCY = register("fuel_efficiency", Codec.FLOAT, ByteBufCodecs.FLOAT, SkillType.FLOAT);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Integer>> FE_CAPACITY = register("fe_capacity", Codec.INT, ByteBufCodecs.INT, SkillType.INT);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Integer>> FE_STORED = register("fe_stored", Codec.INT, ByteBufCodecs.INT);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Boolean>> METAL_GENERATOR = register("metal_generator", Codec.BOOL, ByteBufCodecs.BOOL, SkillType.BOOLEAN);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Boolean>> FOOD_GENERATOR = register("food_generator", Codec.BOOL, ByteBufCodecs.BOOL, SkillType.BOOLEAN);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Boolean>> GEM_GENERATOR = register("gem_generator", Codec.BOOL, ByteBufCodecs.BOOL, SkillType.BOOLEAN);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<GeneratorFuelData>> BURNING_ITEM_DATA = register("burning_item_data", GeneratorFuelData.CODEC, GeneratorFuelData.STREAM_CODEC);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Integer>> PORTABLE_GENERATOR_SLOTS = register("portable_generator_slots", Codec.INT, ByteBufCodecs.INT, SkillType.INT);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Integer>> MAX_BURN_TIME = register("max_burn_time", Codec.INT, ByteBufCodecs.INT);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Integer>> GENERATED_LAST_TICK = register("generated_last_tick", Codec.INT, ByteBufCodecs.INT);
 
     // Utilities
     private static <T> DeferredHolder<DataComponentType<?>, DataComponentType<T>> register(String key, Codec<T> codec, StreamCodec<? super RegistryFriendlyByteBuf, T> streamCodec) {