+++ src/main/java/dev/willyelton/crystal_tools/common/inventory/container/CrystalGeneratorContainerMenu.java
@@ -10,11 +10,11 @@ import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
-public class CrystalGeneratorContainerMenu extends EnergyLevelableContainerMenu {
+public class CrystalGeneratorContainerMenu extends AbstractGeneratorContainerMenu {
     private final CrystalGeneratorBlockEntity blockEntity;
 
     public CrystalGeneratorContainerMenu(int containerId, Level level, BlockPos pos, Inventory playerInventory, ContainerData data) {
-        super(Registration.CRYSTAL_GENERATOR_CONTAINER.get(), containerId, playerInventory, data);
+        super(Registration.CRYSTAL_GENERATOR_CONTAINER.get(), containerId, playerInventory, data, 8, 109);
         blockEntity = (CrystalGeneratorBlockEntity) level.getBlockEntity(pos);
 
         if (blockEntity == null) {
@@ -22,8 +22,6 @@ public class CrystalGeneratorContainerMenu extends EnergyLevelableContainerMenu 
         }
 
         this.addSlot(blockEntity.getFuelHandler(), 0, 80, 59);
-
-        this.layoutPlayerInventorySlots(8, 109);
     }
 
     @Override
@@ -59,35 +57,6 @@ public class CrystalGeneratorContainerMenu extends EnergyLevelableContainerMenu 
         } else {
             return player.distanceToSqr((double) blockEntity.getBlockPos().getX() + 0.5D, (double) blockEntity.getBlockPos().getY() + 0.5D, (double) blockEntity.getBlockPos().getZ() + 0.5D) <= 64.0D;
         }
-    }
-
-    public boolean isLit() {
-        return this.data.get(3) > 0;
-    }
-
-    public float getLitProgress() {
-        if (this.data.get(4) == 0) return 0;
-
-        return this.data.get(3) / (float) this.data.get(4);
-    }
-
-    @Override
-    public float getCurrentEnergy() {
-        return this.data.get(5);
-    }
-
-    @Override
-    public float getMaxEnergy() {
-        return this.data.get(6);
-    }
-
-    public int getCurrentGeneration() {
-        return this.data.get(7);
-    }
-
-    @Override
-    public String getBlockType() {
-        return "generator";
     }
 
     @Override